// Frontend/front/src/app/api/model/paymentMethodEnum.ts
/**
 * CubArt API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const PaymentMethodEnum = {
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3
} as const;
export type PaymentMethodEnum = typeof PaymentMethodEnum[keyof typeof PaymentMethodEnum];